# apps/nestjs/prisma/swagger.yml
openapi: 3.0.0
info:
  title: ft_transcendence API
  description: Welcome to the OpenAPI documentation for the backend of the 'ft_transcendence' project, built with NestJS. This documentation provides a comprehensive overview of the API endpoints available for interacting with the application, including descriptions of each endpoint's functionality, input and output parameters, response codes...
  version: 1.0.0
...
servers:
  - url: http://localhost:3000
    description: NestJS server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /users/me:
    get:
      tags:
        - users
      summary: Get your user information
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              examples:
                base:
                  value: {"id": 1, "username": "new username", "login42": "login42", "avatar": null, "status": "ONLINE", "twoFactorAuth": false, "friends": [], "blocked": [], "blockedBy": [], "friendRequestsSent": [], "friendRequestsReceived": [], "wins": 0, "losses": 0, "history": [], "memberOf": [{"room": {"id": 1, "name": "name", "access": "PUBLIC"}}, {"room": {"id": 3, "name": "name", "access": "PROTECTED"}}, {"room": {"id": 4, "name": "name", "access": "PRIVATE"}}], "createdAt": "2023-05-03T10:57:04.381Z", "updatedAt": "2023-05-03T11:56:56.913Z"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
    put:
      tags:
        - users
      summary: Edit your user information (Does not work with images yet)
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The new username
                  example: New username
                twoFactorAuth:
                  type: string
                  description: Whether to activate the two-factor autentication
                  example: "true"
      responses:
        '200':
          description: User information edited successfully
          content:
            application/json:
              schema:
                type: object
              examples:
                base:
                  value: {"id": 1, "username": "new username", "avatar": null, "twoFactorAuth": false}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              examples:
                empty_username:
                  value: {"statusCode": 400, "message": ["username should not be empty"], "error": "Bad Request"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
    delete:
      tags:
        - users
      summary: Delete your account 
      responses:
        '204':
          description: Account deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
  /users/{id}:
    get:
      tags:
        - users
      summary: Get user information
      parameters:
        - name: id
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              examples:
                base:
                  value: {"id": 1, "username": "username", "login42": "login42", "avatar": null, "status": "ONLINE", "wins": 0, "losses": 0, "history": [], "createdAt": "2023-05-03T10:57:04.381Z"}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              examples:
                invalid_user_id:
                  summary: Invalid user id
                  value: {"statusCode": 400, "message": "Invalid user id", "error": "Bad Request"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              examples:
                user_blocked:
                  summary: User blocked
                  value: {"statusCode": 403, "message": "You have blocked this user", "error": "Forbidden"}
                blocked_by_this_user:
                  summary: Blocked by this user
                  value: {"statusCode": 403, "message": "You have been blocked by this user", "error": "Forbidden"}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
              examples:
                user:
                  summary: User
                  value: {"statusCode": 404, "message": "User not found", "error": "Not found"}
  /users/{id}/block:
    post:
      tags:
        - users
      summary: Block user
      parameters:
        - name: id
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '201':
          description: User blocked successfully
          content:
            application/json:
              schema:
                type: object
              examples:
                base:
                  value: {"userA_id": 1, "userB_id": 2}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              examples:
                invalid_user_id:
                  summary: Invalid user id
                  value: {"statusCode": 400, "message": "Invalid user id", "error": "Bad Request"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              examples:
                self:
                  summary: Self-block
                  value: {"statusCode": 403, "message": "You cannot perform this action on yourself", "error": "Forbidden"}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
              examples:
                user:
                  summary: User
                  value: {"statusCode": 404, "message": "User not found", "error": "Not found"}
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
              examples:
                base:
                  value: {"statusCode": 409, "message": "This user is not blocked", "error":"Conflict"}
    delete:
      tags:
        - users
      summary: Unblock user
      parameters:
        - name: id
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User unblocked successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              examples:
                invalid_user_id:
                  summary: Invalid user id
                  value: {"statusCode": 400, "message": "Invalid user id", "error": "Bad Request"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              examples:
                self:
                  summary: Self-unblock
                  value: {"statusCode": 403, "message": "You cannot perform this action on yourself", "error": "Forbidden"}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
              examples:
                user:
                  summary: User
                  value: {"statusCode": 404, "message": "User not found", "error": "Not found"}
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
              examples:
                base:
                  value: {"statusCode": 409, "message": "This user is already blocked", "error":"Conflict"}
  /users/{id}/friend/request:
    post:
      tags:
        - users
      summary: Send friend request
      parameters:
        - name: id
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Friend request sent successfully
          content:
            application/json:
              schema:
                type: object
              examples:
                friend_request:
                  summary: Friend request
                  value: {"userA_id": 1, "userB_id": 2}
                friend:
                  summary: Friend
                  value: {"userA_id": 1, "userB_id": 2}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              examples:
                invalid_user_id:
                  summary: Invalid user id
                  value: {"statusCode": 400, "message": "Invalid user id", "error": "Bad Request"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              examples:
                self:
                  summary: Send to yourself
                  value: {"statusCode": 403, "message": "You cannot perform this action on yourself", "error": "Forbidden"}
                user_blocked:
                  summary: User blocked
                  value: {"statusCode": 403, "message": "You have blocked this user", "error": "Forbidden"}
                blocked_by_this_user:
                  summary: Blocked by this user
                  value: {"statusCode": 403, "message": "You have been blocked by this user", "error": "Forbidden"}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
              examples:
                user:
                  summary: User
                  value: {"statusCode": 404, "message": "User not found", "error": "Not found"}
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
              examples:
                friend_request_already_sent:
                  summary: Friend request already sent
                  value: {"statusCode": 409, "message": "You have already sent a friend request to this user", "error": "Conflict"}
                already_friends:
                  summary: Already friends
                  value: {"statusCode": 409, "message": "You are already friends with this user", "error": "Conflict"}
    delete:
      tags:
        - users
      summary: Cancel friend request
      parameters:
        - name: id
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Friend request cancelled successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              examples:
                invalid_user_id:
                  summary: Invalid user id
                  value: {"statusCode": 400, "message": "Invalid user id", "error": "Bad Request"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples:
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              examples:
                self:
                  summary: Cancel to yourself
                  value: {"statusCode": 403, "message": "You cannot perform this action on yourself", "error": "Forbidden"}
                user_blocked:
                  summary: User blocked
                  value: {"statusCode": 403, "message": "You have blocked this user", "error": "Forbidden"}
                blocked_by_this_user:
                  summary: Blocked by this user
                  value: {"statusCode": 403, "message": "You have been blocked by this user", "error": "Forbidden"}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
              examples:
                user:
                  summary: User
                  value: {"statusCode": 404, "message": "User not found", "error": "Not found"}
                friend_request:
                  summary: Friend request
                  value: {"statusCode": 404, "message": "Friend request not found", "error": "Not Found"}
  /users/{id}/friend/response:
    post:
      tags:
        - users
      summary: Accept friend request
      parameters:
        - name: id
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Friend request accepted successfully
          content:
            application/json:
              schema:
                type: object
              examples:
                friend:
                  summary: Friend
                  value: {"userA_id": 1, "userB_id": 2}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              examples:
                invalid_user_id:
                  summary: Invalid user id
                  value: {"statusCode": 400, "message": "Invalid user id", "error": "Bad Request"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              examples:
                self:
                  summary: Accept to yourself
                  value: {"statusCode": 403, "message": "You cannot perform this action on yourself", "error": "Forbidden"}
                user_blocked:
                  summary: User blocked
                  value: {"statusCode": 403, "message": "You have blocked this user", "error": "Forbidden"}
                blocked_by_this_user:
                  summary: Blocked by this user
                  value: {"statusCode": 403, "message": "You have been blocked by this user", "error": "Forbidden"}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
              examples:
                user:
                  summary: User
                  value: {"statusCode": 404, "message": "User not found", "error": "Not found"}
                friend_request:
                  summary: Friend request
                  value: {"statusCode": 404, "message": "Friend request not found", "error": "Not found"}
    delete:
      tags:
        - users
      summary: Reject friend request
      parameters:
        - name: id
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Friend request rejected successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              examples:
                invalid_user_id:
                  summary: Invalid user id
                  value: {"statusCode": 400, "message": "Invalid user id", "error": "Bad Request"}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
              examples: 
                base:
                  value: {"statusCode": 401, "message": "Unauthorized"}
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              examples:
                self:
                  summary: Reject to yourself
                  value: {"statusCode": 403, "message": "You cannot perform this action on yourself", "error": "Forbidden"}
                user_blocked:
                  summary: User blocked
                  value: {"statusCode": 403, "message": "You have blocked this user", "error": "Forbidden"}
                blocked_by_this_user:
                  summary: Blocked by this user
                  value: {"statusCode": 403, "message": "You have been blocked by this user", "error": "Forbidden"}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
              examples:
                user:
                  summary: User
                  value: {"statusCode": 404, "message": "User not found", "error": "Not found"}
                friend_request:
                  summary: Friend request
                  value: {"statusCode": 404, "message": "Friend request not found", "error": "Not found"}
...